seperating code for a bit more readability
